Assignment 3/Studio/requirements.txt
langgraph
langgraph-prebuilt
langchain-core
langchain-community
langchain-openai
